Chapter 6 - Parsing
Still not happy with the Java-ish-ness of this code. It could be
improved by separating the buffer of tokens from the "parser". Right
now, parsing is just functions that advance the position in the buffer
depending on whether the desired token matches.

There's a lot of confusion around advance/peek/previous/etc that
just does not read well. Direct indexing or iterators (peekable?) would
be more Rust-like, if it can be achieved.

I'm not happy with the discriminant stuff on TokenType either. It's
very much a compromise to cut down the awkwardness of keeping literals
separately from the token type in the original Java code. The derive_more
crate could help here with its IsVariant derive.