Chapter 7 - Evaluating Expressions
Decided not to give into the idea that there should be a struct for the interpreter,
since we are only evaluating expressions. The code turned out pretty nice, and I got
to make use of some stdlib traits to make the code more ergonomic.

It occurs to me that there might be an impl of From<Result<T,E>> for Result<U,E> where
U: From<T>. However, I didn't see it immediately in the standard library, so we have
Ok(something.into()) all over the evaluate function. It's ok.

I think the TypeError variant came out pretty well and I intuited it would be needed
before the author brought it up. Doing this in Rust where I have to specify a variant
for the value types forced my hand.